Additional html files for "zoovy-web-existing-content"
Extracted all content from existing Zoovy site into html files. Those
files are in a folder called "zoovy-web-existing-content".

I left all of the files/ folders that weren't html files. Please go
through and delete what is not needed.

I have a copy of all of the files from the existing site as it was
yesterday on my computer.